Add initial decorator work to Python API
Some of the people who have looked at the Python API have said that it
isn't very Pythonic. Mostly they've pointed to the fact that the user
is required to create classes for things that don't have any state,
such as computations. In order to address this I've added some
decorator functions that reduce the number of classes that the user
needs to create.

In most cases the decorators just return instances of the classes that
the user would have written originally, so the changes to the API are
minimal. However, I've gotten rid of the need for a state builder by
adding a "build" class method to the state class; now the state class
is effectively the original state class AND the builder class, but the
user no longer has to explicitly write the builder. We could probably
get rid of the builder idea altogether by passing the state class to
the application build and having Wallaroo call the class's constructor
when it needs to build one.

The only example that has been updated is the WordCount example, but
it should be possible to get a feel for the kinds of changes that I
think should be made by looking at this example.

This is a work in progress meant to address issue #1604.